group similar settings in preparation of tabular settings page
