+++ static/app-strings.json
@@ -1269,5 +1269,9 @@
   "These changes will appear shortly.": "These changes will appear shortly.",
   "Block Channel": "Block Channel",
   "Manage Notifications": "Manage Notifications",
+  "Uploads": "Uploads",
+  "Advanced Settings": "Advanced Settings",
+  "Browse": "Browse",
+  "Automatic Transcoding": "Automatic Transcoding",
   "--end--": "--end--"
 }
